nutrition_insights/phase3/requirements.txt
streamlit
pandas
numpy
requests
feedparser
python-dotenv
praw
beautifulsoup4
python-dateutil